fixed issues from reviewer's comments
