+++ src/components/navigation/ProfilePopover.tsx
@@ -24,12 +24,10 @@ export function ProfilePopover() {
     <Popover>
       <PopoverTrigger asChild>
         <Button variant="ghost" className="relative h-9 w-9 rounded-full p-0">
-         {userImage != null && (
-            <Avatar>
-              <AvatarImage src={userImage} alt={userName} />
-              <AvatarFallback>{userInitials}</AvatarFallback>
-            </Avatar>
-         )}
+          <Avatar>
+            <AvatarImage src={userImage} alt={userName} />
+            <AvatarFallback>{userInitials}</AvatarFallback>
+          </Avatar>
         </Button>
       </PopoverTrigger>
       <PopoverContent className="w-56 flex flex-col gap-2">
