+++ src/domain/hooks/meetingHooks.ts
@@ -4,6 +4,9 @@ import { useQuery } from '@tanstack/react-query';
 
 import { MeetingData } from '@/domain/entities/calendar/MeetingData';
 import { meetingRepository } from '@/repositories/meetingRepository';
+import { DragDropContext, DragDropProvider } from '@/components/MonthViewCalendar/DragDropContex';
+import { useContext } from 'react';
+
 
 export const useMeetings = (enabled = true) =>
   useQuery<MeetingData[]>({
@@ -11,3 +14,11 @@ export const useMeetings = (enabled = true) =>
     queryFn: meetingRepository.getMeetings,
     enabled,
   });
+
+  export function useDragDrop() {
+  const context = useContext(DragDropContext);
+  if (context === undefined) {
+    throw new Error('useDragDrop must be used within a DragDropProvider');
+  }
+  return context;
+}
